fix rsqrt test
