+++ paddle/phi/backends/xpu/xpu2_op_list.cc
@@ -777,13 +777,9 @@ XPUOpMap& get_kl2_ops() {
                      phi::DataType::INT64,
                      phi::DataType::FLOAT16})},
       {"reduce_mean_grad",
-       XPUKernelSet({phi::DataType::FLOAT32,
-                     phi::DataType::FLOAT16,
-                     phi::DataType::BFLOAT16})},
+       XPUKernelSet({phi::DataType::FLOAT32, phi::DataType::FLOAT16})},
       {"reduce_mean",
-       XPUKernelSet({phi::DataType::FLOAT32,
-                     phi::DataType::FLOAT16,
-                     phi::DataType::BFLOAT16})},
+       XPUKernelSet({phi::DataType::FLOAT32, phi::DataType::FLOAT16})},
       {"reduce_min_grad", XPUKernelSet({phi::DataType::FLOAT32})},
       {"reduce_min",
        XPUKernelSet({phi::DataType::FLOAT32,
